// src/app/_interfaces/categoria.interface.ts
export interface ICategoria {
    id: number;
    categoryid: number;
    storeid: number;
    name: string;
    image: string;
    status: number;
    createdBy: number;
    createdDT: string; // 2020-04-06 15:54:01,
    modifiedBy: number;
    modifiedDT: string; //2021-05-25 17:45:24,
    parent: string;
    parentimage: string; //img/category/6_1600457280.ico,
    sid: number;
    sname: string;
}